% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{AFM}
\alias{AFM}
\title{AFM data}
\description{
1104 calc-alkaline and tholeiitic basalt compositions
}
\examples{
data(AFM,package='geostats')
ternary(AFM[,-1],pch=21,bg=AFM[,1])
}
\keyword{data}
