% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractals.R
\name{boxcount}
\alias{boxcount}
\title{box counting}
\usage{
boxcount(mat, size)
}
\arguments{
\item{mat}{a square square matrix of 0s and 1s, whose size should
be a power of 2.}

\item{size}{the size (pixels per side) of the boxes, whose size
should be a power of 2.}
}
\value{
an integer
}
\description{
Count the number of boxes needed to cover all the 1s
    in a matrix of 0s and 1s.
}
\examples{
g <- sierpinski(n=5)
boxcount(mat=g,size=16)
}
