% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{catchments}
\alias{catchments}
\title{properties of 20 river catchments}
\description{
six different (three discrete, three continuous) measurements for
twenty fictitious river catchments, containing their dominant
lithology (categorical data), stratigraphic age (ordinal data),
number of springs (count data), the pH of the river water
(Cartesian quantity), the area of the catchment in km\eqn{^2}
(Jeffreys quantity) and the percentage covered by vegetation
(proportion).
}
\examples{
data(catchments,package='geostats')
hist(catchments$pH)
}
\keyword{data}
