% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional.R
\name{circle.plot}
\alias{circle.plot}
\title{plot circular data}
\usage{
circle.plot(a, degrees = FALSE, tl = 0.1, ...)
}
\arguments{
\item{a}{angle(s), scalar or vector}

\item{degrees}{logical. \code{TRUE} for degrees, \code{FALSE} for
radians}

\item{tl}{tick length (value between 0 and 1)}

\item{...}{optional arguments to be passed on to the generic
\code{matlines} function}
}
\value{
no return value
}
\description{
Plots directional data as ticks on a circle, with
    angles plotting in a clockwise direction from the top.
}
\examples{
data(striations,package='geostats')
circle.plot(striations,degrees=TRUE)
}
