% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional.R
\name{circle.points}
\alias{circle.points}
\title{add points to a circular plot}
\usage{
circle.points(angles, degrees = FALSE, ...)
}
\arguments{
\item{angles}{scalar or vector}

\item{degrees}{\code{TRUE} for degrees, \code{FALSE} for radians}

\item{...}{optional arguments to be passed on to the generic
\code{points} function}
}
\description{
adds directional data as points on an existing circle
    plot
}
\details{
adds points to a circle with angles plotting in a
    clockwise direction from the top
}
\examples{
data(striations,package='geostats')
circle.plot(striations,degrees=TRUE)
md <- meanangle(striations,degrees=TRUE)
circle.points(md,pch=22,bg='black')
}
