% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{clasts}
\alias{clasts}
\title{clast size data}
\description{
20 clast size measurements, in cm
}
\examples{
data(clasts,package='geostats')
d <- density(log(clasts))
plot(d)
}
\keyword{data}
