% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{declustered}
\alias{declustered}
\title{declustered earthquake data}
\description{
dataset of 28267 earthquakes between 1769 and 2016, with
aftershocks and precursor events removed
}
\examples{
data(declustered,package='geostats')
quakesperyear <- countQuakes(declustered,minmag=5.0,from=1917,to=2016)
table(quakesperyear)
}
\references{
Mueller, C.S., 2019. Earthquake catalogs for the USGS
    national seismic hazard maps. Seismological Research Letters,
    90(1), pp.251-261.
}
\keyword{data}
