% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{DZ}
\alias{DZ}
\title{detrital zircon U-Pb data}
\description{
Detrital zircon U-Pb data of 13 sand samples from China.
}
\examples{
data(DZ,package='geostats')
qqplot(DZ[['Y']],DZ[['5']])
}
\references{
Vermeesch, P. ``Multi-sample comparison of detrital age
    distributions.''  Chemical Geology 341 (2013): 140-146.
}
\keyword{data}
