% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{earthquakes}
\alias{earthquakes}
\title{earthquake data}
\description{
Dataset of 20000 earthquakes between 2017 and 2000, downloaded from
the USGS earthquake database
(\url{https://earthquake.usgs.gov/earthquakes/search/}).
}
\examples{
data(earthquakes,package='geostats')
gutenberg(earthquakes$mag)
}
\keyword{data}
