% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{FAM}
\alias{FAM}
\title{A-F-M data}
\description{
FeO - (Na\eqn{_2}O + K\eqn{_2}O) - MgO compositions of 630
calc-alkali basalts from the Cascade Mountains and 474 tholeiitic
basalts from Iceland. Arranged in F-A-M order instead of A-F-M for
consistency with the \code{ternary} function.
}
\examples{
data(FAM,package='geostats')
ternary(FAM[,-1])
}
\keyword{data}
