% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{fault}
\alias{fault}
\title{fault orientation data}
\description{
Ten paired strike and dip measurements (in degrees), drawn from a
von Mises - Fisher distribution with mean vector
\eqn{\mu=\{1,1,1\}/\sqrt{3}} and concentration parameter
\eqn{\kappa=200}.
}
\examples{
data(fault,package='geostats')
stereonet(fault,option=2,degrees=TRUE,show.grid=FALSE)
}
\keyword{data}
