% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{forams}
\alias{forams}
\title{foram count data}
\description{
Planktic foraminifera counts in surface sediments in the Atlantic ocean.
}
\examples{
data(forams,package='geostats')
abundant <- forams[,c('quinqueloba','pachyderma','incompta',
                      'glutinata','bulloides')]
other <- rowSums(forams[,c('uvula','scitula')])
dat <- cbind(abundant,other)
chisq.test(dat)
}
\keyword{data}
