% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractals.R
\name{fractaldim}
\alias{fractaldim}
\title{calculate the fractal dimension}
\usage{
fractaldim(mat, plot = TRUE, ...)
}
\arguments{
\item{mat}{a square matrix of 0s and 1s. Size must be a power of 2.}

\item{plot}{logical. If \code{TRUE}, plots the results on a log-log
scale.}

\item{...}{optional arguments to the generic \code{points} function.}
}
\description{
performs box counting on a matrix of 0s and 1s.
}
\examples{
g <- sierpinski(n=5)
fractaldim(g)
}
