% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostats.R
\docType{package}
\name{geostats}
\alias{geostats}
\alias{_PACKAGE}
\alias{geostats-package}
\title{library(geostats)}
\description{
A list of documented functions may be viewed by typing
    \code{help(package='geostats')}.  Detailed instructions are
    provided at \url{https://github.com/pvermees/geostats/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pvermees/geostats/}
}

}
\author{
\strong{Maintainer}: Pieter Vermeesch \email{p.vermeesch@ucl.ac.uk}

}
