% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{islands}
\alias{islands}
\title{properties of 20 islands}
\description{
six different (three discrete, three continuous) measurements for
twenty fictitious islands, containing their dominant lithology
(categorical data), stratigraphic age (ordinal data), number of
springs (count data), the pH of the spring water (Cartesian
quantity), the area of the island in km\eqn{^2} (Jeffreys quantity)
and the percentage covered by vegetation (proportion).
}
\examples{
data(islands,package='geostats')
hist(islands$pH)
}
\keyword{data}
