% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractals.R
\name{koch}
\alias{koch}
\title{Koch snowflake}
\usage{
koch(n = 4, plot = TRUE, res = 512)
}
\arguments{
\item{n}{an integer value controling the number of recursive
levels.}

\item{plot}{logical.  If \code{TRUE}, the Koch flake is plotted.}

\item{res}{the number of pixels in each side of the output matrix}
}
\value{
a \code{res x res} matrix with 0s and 1s
}
\description{
Calculates or plots a Koch set of fractal lines.
}
\details{
The Koch set is generated using a recursive algorithm that
    is built on a triangular hat shaped line segment. Each level of
    recursion replaces each linear segment by the same pattern.
}
\examples{
k <- koch(n=5)
d <- fractaldim(k,plot=FALSE)
print(d)
}
