% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{palaeomag}
\alias{palaeomag}
\title{palaeomagnetic data}
\description{
Ten paired magnetic declination (azimuth) and inclination (dip)
measurements, drawn from a von Mises - Fisher distribution with
mean vector \eqn{\mu=\{2,2,1\}/3} and concentration parameter
\eqn{\kappa=200}.
}
\examples{
data(palaeomag,package='geostats')
stereonet(trd=palaeomag$decl,plg=palaeomag$incl,degrees=TRUE,show.grid=FALSE)
}
\keyword{data}
