% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{pebbles}
\alias{pebbles}
\title{pebble orientations}
\description{
Orientations (in degrees) of 20 pebbles.
}
\examples{
data(pebbles,package='geostats')
circle.plot(pebbles,degrees=TRUE)
m <- meanangle(pebbles,option=0,orientation=TRUE)
circle.points(m,degrees=TRUE,pch=22,bg='white')
}
\keyword{data}
