% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractals.R
\name{pendulum}
\alias{pendulum}
\title{3-magnet pendulum experiment}
\usage{
pendulum(
  startpos = c(-2, 2),
  startvel = c(0, 0),
  src = rbind(c(0, 0), c(0.5, sqrt(0.75)), c(1, 0)),
  plot = TRUE
)
}
\arguments{
\item{startpos}{2-element vecotor with the initial position}

\item{startvel}{2-element vector with the initial velocity}

\item{src}{n x 2 matrix with the positions of the magnets}

\item{plot}{logical. If \code{TRUE}, generates a plot with the
trajectory of the pendulum.}
}
\value{
the end position of the pendulum
}
\description{
simulate the 3-magnet pendulum experiment
}
\details{
start a pendulumn at a specified position and with a start
    velocity.
}
\examples{
par(mfrow=c(1,2))
pendulum(startpos=c(2,2))
pendulum(startpos=c(1.9,2))
}
