% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{pH}
\alias{pH}
\title{pH data}
\description{
pH measurements in 20 samples of rain water

20 clast size measurements, in cm
}
\examples{
data(pH,package='geostats')
hist(pH)
data(clasts,package='geostats')
d <- density(log(clasts))
plot(d)
}
\keyword{data}
