% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{pop}
\alias{pop}
\title{pop}
\description{
The world population from 1750 until 2014
}
\examples{
data(pop,package='geostats')
plot(population ~ year, data=pop, type='b')
}
\keyword{data}
