% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{porosity}
\alias{porosity}
\title{porosity data}
\description{
20 porosity measurements, as fractions.
}
\examples{
data(porosity,package='geostats')
plot(density(logit(porosity)))
}
\keyword{data}
