% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{randy}
\alias{randy}
\title{generate bivariate random data}
\usage{
randy(pop = 1, n = 250)
}
\arguments{
\item{pop}{an integer from 1 to 4 marking the population of choice:
1 = circle, 2 = arrow, 3 = solid square, 4 = ellipse.}

\item{n}{the number of random draws to be drawn from population \code{pop}}
}
\value{
a \code{[2xn]} matrix of random numbers
}
\description{
Returns bivariate datasets from four synthetic
    distributions that have the shape of a circle, arrow, square
    and ellipse.
}
\examples{
p <- par(mfrow=c(1,4))
for (i in 1:4){
   plot(randy(pop=i))
}
par(p)
}
