% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional.R
\name{Rbar2kappa}
\alias{Rbar2kappa}
\title{\eqn{\bar{R}} to \eqn{\kappa} conversion}
\usage{
Rbar2kappa(R)
}
\arguments{
\item{R}{a scalar or vector of values between 0 and 1}
}
\value{
value(s) between 0 and \eqn{+\infty}
}
\description{
converts concentration parameter \eqn{\bar{R}} to
    \eqn{\kappa}
}
\details{
\eqn{\bar{R}} and \eqn{\kappa} are two types of
    concentration parameter that are commonly used in directional
    data analysis.  \eqn{\kappa} is one of the parameters of the
    parametric von Mises distribution, which is difficult to
    estimate from the data. \eqn{\bar{R}} is easier to calculate
    from data. \code{R2kappa} converts \eqn{\bar{R}} to
    \eqn{\bar{\kappa}} using a lookup table.
}
\examples{
data(striations,package='geostats')
Rbar2kappa(Rbar(striations,degrees=TRUE))
}
