% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{rbsr}
\alias{rbsr}
\title{Rb-Sr data}
\description{
synthetic dataset of 8 Rb-Sr analysis that form a 1Ga isochron
}
\examples{
data(rbsr,package='geostats')
plot(rbsr[,'RbSr'],rbsr[,'SrSr'])
fit <- lm(SrSr ~ RbSr,data=rbsr)
abline(fit)
}
\keyword{data}
