% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{semivariogram}
\alias{semivariogram}
\title{semivariogram}
\usage{
semivariogram(x, y, z, bw = NULL, plot = TRUE, fit = TRUE, ...)
}
\arguments{
\item{x}{numerical vector}

\item{y}{numerical vector of the same length as \code{x}}

\item{z}{numerical vector of the same length as \code{x}}

\item{bw}{(optional) the bin width of the semivariance search
algorithm}

\item{plot}{logical. If \code{FALSE}, suppresses the graphical
output}

\item{fit}{logical. If \code{TRUE}, returns the sill, nugget and
range.}

\item{...}{optional arguments to be passed on to the generic
\code{plot} function}
}
\value{
if \code{fit=TRUE}, returns a vector with the sill, nugget
    and range. If \code{fit=FALSE}, returns the estimated
    semivariances at different distances for the data.
}
\description{
computes, plots, and fits the semivariogram of spatial
    data
}
\details{
Plots the semivariance of spatial data against
    inter-sample distance, and fits a spherical equation to it.
}
\examples{
data(meuse,package='geostats')
semivariogram(x=meuse$x,y=meuse$y,z=log(meuse$cadmium))
}
