% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{skew}
\alias{skew}
\title{calculate the skewness of a dataset}
\usage{
skew(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a scalar
}
\description{
Compute the third moment of a sampling distribution.
}
\examples{
data(catchments,package='geostats')
skew(catchments$vegetation)
}
