% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereonet.R
\name{stereonet}
\alias{stereonet}
\title{stereonet}
\usage{
stereonet(
  azimuth,
  dip,
  wulff = TRUE,
  degrees = FALSE,
  show.lines = TRUE,
  tl = 0.05,
  ...
)
}
\arguments{
\item{azimuth}{angle, in degrees, between 0 and 360 (if
\code{degrees=TRUE}) or between 0 and \eqn{2\pi} (if
\code{degrees=FALSE}).}

\item{dip}{angle, in degrees, between 0 and 90 (if
\code{degrees=TRUE}) or between 0 and \eqn{2\pi} (if
\code{degrees=FALSE}).}

\item{wulff}{logical. If \code{FALSE}, produces a Schmidt net.}

\item{degrees}{logical. If \code{FALSE}, assumes that
\code{azimuth} and \code{dip} are in radians.}

\item{show.lines}{logical. If \code{TRUE}, decorates the plot with
a grid of great and small circles.}

\item{tl}{tick length for the N, E, S, W markers (value between 0
and 1).  Set to 0 to omit the markers.}

\item{...}{optional arguments to be passed on to the generic
\code{points} function}
}
\description{
Plots directional data on a Schmidt equal area or
    Wulff equal angle stereonet.
}
\details{
The Schmidt equal area polar Lambert projection is used to
    plot structural data. Given a set of azimuth (\eqn{a}) and dip
    (\eqn{d}) measurements, it plots the data at positions

\eqn{x = \sqrt{2}\sin(\pi\frac{90-d}{360})
                 \cos(\pi\frac{90-a}{180})}

and

\eqn{y = \sqrt{2}\sin(\pi\frac{90-d}{360})
                 \sin(\pi\frac{90-a}{180})}

The Wulff equal angle polar Lambert projection is used to plot
    crystallographic data. Given a set of azimuth (\eqn{a}) and dip
    (\eqn{d}) measurements, it plots the data at positions

\eqn{x = \tan(\pi\frac{90-d}{360})
         \cos(\pi\frac{90-a}{180})}

and

\eqn{y = \tan(\pi\frac{90-d}{360})
         \sin(\pi\frac{90-a}{180})}
}
\examples{
stereonet(azimuth=c(120,80),dip=c(10,30),degrees=TRUE)
}
\author{
based on the Matlab script by Gerry Middleton
}
