% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{striations}
\alias{striations}
\title{directions of glacial striations}
\description{
Directions (in degrees) of 30 glacial striation measurements from
Madagascar.
}
\examples{
data(striations,package='geostats')
circle.plot(striations,degrees=TRUE)
}
\keyword{data}
