% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{test}
\alias{test}
\title{composition of oceanic basalts}
\description{
major element compositions of 64 island arc basalts (IAB), 23 mid
oceanic ridge basalts (MORB) and 60 ocean island basalts
(OIB). This dataset can be used to test supervised learning
algorithms.
}
\examples{
library(MASS)
data(training,package='geostats')
qd <- qda(affinity ~ ., data=training)
pr <- predict(qd,newdata=test[,-1])
table(test$affinity,pr$class)
}
\keyword{data}
