% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{training}
\alias{training}
\title{composition of oceanic basalts}
\description{
major element compositions of 227 island arc basalts (IAB), 221 mid
oceanic ridge basalts (MORB) and 198 ocean island basalts
(OIB). This dataset can be used to train supervised learning
algorithms.
}
\examples{
library(MASS)
data(training,package='geostats')
qd <- qda(affinity ~ ., data=training)
pr <- predict(qd)
table(training$affinity,pr$class)
}
\keyword{data}
