% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{worldpop}
\alias{worldpop}
\title{world population}
\description{
The world population from 1750 until 2014
}
\examples{
data(worldpop,package='geostats')
plot(worldpop)
}
\keyword{data}
