% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositions.R
\name{xyz2xy}
\alias{xyz2xy}
\title{get x,y plot coordinates of ternary data}
\usage{
xyz2xy(xyz)
}
\arguments{
\item{xyz}{an \code{n x 3} matrix or data frame}
}
\value{
an \code{n x 2} numerical matrix
}
\description{
helper function to generate bivariate plot coordinates
    for ternary data
}
\examples{
xyz <- rbind(c(1,0,0),c(0,1,0),c(0,0,1),c(1,0,0))
xy <- xyz2xy(xyz)
plot(xy,type='l',bty='n')
}
