% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{AFM}
\alias{AFM}
\title{A-F-M data}
\description{
(Na\eqn{_2}O + K\eqn{_2}O) -- FeO - MgO compositions of 630
calc-alkali basalts from the Cascade Mountains and 474 tholeiitic
basalts from Iceland.
}
\examples{
data(AFM,package='geostats')
ternary(AFM[,-1])
}
\keyword{data}
