% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{Britain}
\alias{Britain}
\title{British coast}
\description{
a 512 x 512 pixel image of the British coast line
}
\examples{
data(Britain,package='geostats')
p <- par(mfrow=c(1,2))
image(Britain)
fractaldim(Britain)
}
\keyword{data}
