% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional.R
\name{circle.plot}
\alias{circle.plot}
\title{plot circular data}
\usage{
circle.plot(angles, degrees = FALSE, tl = 0.1, ...)
}
\arguments{
\item{angles}{scalar or vector}

\item{degrees}{\code{TRUE} for degrees, \code{FALSE} for radians}

\item{tl}{tick length (value between 0 and 1)}

\item{...}{optional arguments to be passed on to the generic
\code{matlines} function}
}
\description{
Plots directional data as ticks on a circle
}
\details{
Produces a circle with angles plotting in a clockwise
    direction from the top
}
\examples{
data(striations,package='geostats')
circle.plot(striations,degrees=TRUE)
}
