% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{colourplot}
\alias{colourplot}
\title{colour plot}
\usage{
colourplot(
  x,
  y,
  z,
  X,
  Y,
  Z,
  levels,
  nlevels = 20,
  colspec = rainbow,
  pch = 21,
  cex = 1,
  plot.title,
  plot.axes,
  key.title,
  key.axes,
  asp = NA,
  xaxs = "i",
  yaxs = "i",
  las = 1,
  axes = TRUE,
  frame.plot = axes,
  extra,
  ...
)
}
\arguments{
\item{x}{numerical vector of \eqn{n} equally spaced values}

\item{y}{numerical vector of \eqn{m} equally spaced values}

\item{z}{an \eqn{n \times m} matrix of numerical values}

\item{X}{numerical vector of \eqn{N} values}

\item{Y}{numerical vector of \eqn{N} values}

\item{Z}{numerical vector of \eqn{N} values}

\item{levels}{a set of levels which are used to partition the range
of \code{z}. Must be *strictly* increasing (and finite).  Areas
with \code{z} values between consecutive levels are painted
with the same colour.}

\item{nlevels}{if \code{levels} is not specified, the range of
\code{z}, values is divided into approximately this many
levels.}

\item{colspec}{colour specification (e.g., \code{rainbow},
\code{hsv}, \code{hcl}, \code{rgb})}

\item{pch}{plot character (21 - 25)}

\item{cex}{plot character magnification}

\item{plot.title}{statements that add titles to the main plot.}

\item{plot.axes}{statements that draw axes on the main plot. This
overrides the default axes.}

\item{key.title}{statements that add titles for the plot key.}

\item{key.axes}{statements that draw axes on the plot key.  This
overrides the default axis.}

\item{asp}{the y/x aspect ratio, see \code{\link{plot.window}}.}

\item{xaxs}{the x axis style.  The default is to use internal
labeling.}

\item{yaxs}{the y axis style.  The default is to use internal
labeling.}

\item{las}{the style of labeling to be used.  The default is to use
horizontal labeling.}

\item{axes}{logicals indicating if axes should be drawn}

\item{frame.plot}{logicals indicating if a box should be drawn, as
in \code{\link{plot.default}}.}

\item{extra}{(optional) extra intructions to be carried out in the
main plot window, such as text annotations.}

\item{...}{additional graphical parameters}
}
\description{
combines a filled contour plot and filled scatter plot
    for 3-dimensional measurements
}
\details{
adds a colour bar to a scatter plot and/or filled contour
    plot.  This funciton, which is based on base \code{R}'s
    \code{filled.contour} function, is useful for visualising
    kriging results.
}
\examples{
data('meuse',package='geostats')
colourplot(X=meuse$x,Y=meuse$y,Z=log(meuse$zinc))
}
