% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{Corsica}
\alias{Corsica}
\title{rivers on Corsica}
\description{
a 512 x 512 pixel image of the river network on Corsica
}
\examples{
data(Corsica,package='geostats')
p <- par(mfrow=c(1,2))
image(Corsica)
fractaldim(Corsica)
}
\keyword{data}
