% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{DZ}
\alias{DZ}
\title{detrital zircon U-Pb data}
\description{
detrital zircon U-Pb data of 5 sand samples from China
}
\examples{
data(DZ,package='geostats')
qqplot(DZ[['Y']],DZ[['5']])
}
\keyword{data}
