% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{fractures}
\alias{fractures}
\title{fractures}
\description{
A \eqn{512 \times 512} pixel image of a fracture network.
}
\examples{
data(fractures,package='geostats')
p <- par(mfrow=c(1,2))
image(fractures)
fractaldim(fractures)
par(p)
}
\keyword{data}
