% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{hills}
\alias{hills}
\title{hills}
\description{
This data set contains 150 X-Y-Z values for a synthetic landscape
that consists of three Gaussian mountains.
}
\examples{
data(hills,package='geostats')
semivariogram(x=hills$X,y=hills$Y,z=hills$Z)
}
\keyword{data}
