% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{ksdist}
\alias{ksdist}
\title{Kolmogorov-Smirnov distance matrix}
\usage{
ksdist(dat)
}
\arguments{
\item{dat}{a list of numerical data vectors}
}
\description{
fills a square matrix with Kolmogorov-Smirnov
    statistics
}
\examples{
data(DZ,package='geostats')
d <- ksdist(DZ)
plot(cmdscale(d))
}
