% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{major}
\alias{major}
\title{composition of Namib dune sand}
\description{
major element compositions of 16 Namib sand samples
}
\examples{
data(major,package='geostats')
comp <- clr(major)
pc <- prcomp(comp)
biplot(pc)
}
\keyword{data}
