% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional.R
\name{meanangle}
\alias{meanangle}
\title{von Mises distribution}
\usage{
meanangle(a, degrees = FALSE)
}
\arguments{
\item{a}{angle(s), scalar or vector}

\item{degrees}{\code{TRUE} for degrees, \code{FALSE} for radians}
}
\value{
the mean angle, either in radians (if
    \code{degrees=FALSE}), or in degrees.
}
\description{
returns the probability density of a von Mises distribution
}
\details{
the von Mises distribution describes probability distributions
on a circle using the following density function:

\eqn{\frac{\exp(\kappa\cos(x-\mu))}{2\pi I_0(\kappa)}}

where \eqn{I_0(\kappa)} is a zero order Bessel function
}
\examples{
data(striations,package='geostats')
circle.plot(a=striations,degrees=TRUE)
circle.points(meanangle(striations,degrees=TRUE),pch=19)
}
