% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{PCA2D}
\alias{PCA2D}
\title{Principal Component Analysis of 2D data}
\usage{
PCA2D(X)
}
\arguments{
\item{X}{a matrix with two columns}
}
\description{
produces a 4-panel summary plot for two dimensional
    PCA for didactical purposes
}
\examples{
X <- rbind(c(-1,7),c(3,2),c(4,3))
colnames(X) <- c('a','b')
PCA2D(X)
}
