% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional.R
\name{Rbar}
\alias{Rbar}
\title{calculate \eqn{\bar{R}}}
\usage{
Rbar(a, degrees = FALSE)
}
\arguments{
\item{a}{vector of angles}

\item{degrees}{\code{TRUE} for degrees, \code{FALSE} for radians}
}
\value{
a value between 0 and 1
}
\description{
returns \eqn{\bar{R}}, a measure of directional concentration
}
\details{
Given \eqn{n} directional measurements \eqn{\theta_i},

\eqn{
\bar{R} =
\sqrt{\frac{\sum_{i=1}^{n}(\sin(\theta_i)^2 + cos(\theta_i)^2))}{n} }
}
}
\examples{
data(striations,package='geostats')
Rbar(a=striations,degrees=TRUE)
}
